requirements.txt
accelerate==1.8.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
clickhouse-connect==0.8.18
colorama==0.4.6
colpali_engine @ git+https://github.com/illuin-tech/colpali.git@09eafbf50b967913e5c8088ff6bea076f33bac5c
distro==1.9.0
dnspython==2.7.0
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.67.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
loguru==0.7.3
lxml==6.0.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.45.0
networkx==3.5
nomic==3.5.3
numpy==2.3.1
openai==1.93.0
packaging==24.2
pandas==2.3.0
peft==0.15.2
pillow==11.3.0
protobuf==3.20.3
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydotenv==0.0.7
Pygments==2.19.2
PyJWT==2.10.1
pymilvus==2.5.11
pymongo==4.13.2
Pympler==1.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
replicate==1.0.7
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
safetensors==0.5.3
scipy==1.16.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.40.0
streamlit-drawable-canvas==0.9.3
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.21.2
toml==0.10.2
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
validators==0.35.0
watchdog==5.0.3
win32_setctime==1.2.0
zipp==3.23.0
zstandard==0.23.0
